Update HammingI.py
This includes all of the field.math stuff at the top bc i couldn't figure out how to import it/not have it defined in the same file.  Trying to use field math to create the encoded message is currently returning datatype error and I haven't been able to correct it. I am still working on the parity check and error correction portion of the code, but am focusing more on write-up right now.